// src/app/api/swagger/models/api-market-participant-offer-info-dto.ts
export interface ApiMarketParticipantOfferInfoDto {
  Actions?: ApiParticipantOfferActionsDto | null;
  BarCode?: string | null;
  Category?: ApiB2BCategoryDto | null;
  /** @deprecated */CharacteristicValues?: Array<ApiCharacteristicValuesDto> | null;
  Characteristics?: Array<ApiCharacteristicDto> | null;
  ClassificatorCode?: string | null;
  ClassificatorDescription?: string | null;
  Country?: ApiCountryDto | null;
  CountryCode?: string | null;
  CountryOfOrigin?: string | null;
  CreatedMethod?: ApiOfferCreatedMethodEnum;
  CurrencyCode?: string | null;
  DeliveryCost?: number | null;
  DeliveryPlace?: string | null;
  DeliveryRegionInfos?: Array<ApiDeliveryRegionInfoDto> | null;
  DeliveryRegions?: string | null;
  DeliveryTerms?: string | null;
  DisplayCharacteristics?: Array<ApiDisplayCharacteristicDto> | null;
  DosageForm?: string | null;
  DosageMeasureUnits?: string | null;
  DrugDosage?: string | null;
  DuringDays?: number | null;
  ExclusiveToTradeId?: number | null;
  ExclusiveToTradeNumber?: string | null;
  ExternalSystemIdentifier?: string | null;
  ExternalSystemIdentifierNumeric?: number | null;
  Id?: number;
  IsFromGuarnteeSupplier?: boolean;
  IsOwner?: boolean;
  IsParticipatesInRequirementRequests?: boolean;
  IsPhysicalPerson?: boolean;
  IsPublicOffer?: boolean;
  KsrCode?: string | null;
  KsrName?: string | null;
  KsrUnit?: string | null;
  Manufacturer?: string | null;
  ManufacturerCountryRegistryNumber?: string | null;
  ManufacturerCountryRegistryType?: ApiManufacturerCountryRegistryTypeEnum | null;
  ManufacturerCountryScore?: number | null;
  MovingProLink?: string | null;
  OfferOwner?: boolean;
  OfferState?: ApiParticipantOfferStateEnum;
  OfferStateDescription?: string | null;
  OfferTerms?: string | null;
  OfferType?: ApiParticipantOfferTypeEnum;
  OfferTypeDescription?: string | null;
  OkeiCode?: string | null;
  OkeiDescription?: string | null;
  Okpd2Values?: Array<ApiOkpd2MarketDto> | null;
  Params?: Array<ApiYmlParamDto> | null;
  ParticipantCountry?: ApiOrganizationCountryAltEnum;
  ParticipantForeignId?: string | null;
  ParticipantGuid?: string;
  ParticipantId?: number;
  ParticipantInn?: string | null;
  ParticipantKpp?: string | null;
  ParticipantName?: string | null;
  ParticipantOfferDocuments?: Array<ApiParticipantOfferDocumentDto> | null;
  ParticipantOfferSource?: ApiParticipantOfferSourceEnum;
  ParticipantOfferSpecific?: ApiParticipantOfferSpecificFlagEnum | null;
  ParticipantSource?: number | null;
  PossibleDeliveryBySeller?: boolean;
  PossibleSelfDelivery?: boolean | null;
  PossibleSendByTransportCompany?: boolean;
  PriceListDeliveryRegionInfos?: Array<ApiDeliveryRegionInfoDto> | null;
  PriceListId?: number | null;
  PriceListState?: ApiPriceListStateEnum;
  PriceListType?: ApiPriceListTypeEnum;
  PriceRegionDetails?: Array<ApiPriceRegionDetailDto> | null;
  PrintFormFileName?: string | null;
  PrintFormGuid?: string | null;
  Producer?: string | null;
  ProductDescription?: string | null;
  ProductMNN?: string | null;
  ProductName?: string | null;
  PublishDate?: string | null;
  PublishDateStart?: string | null;
  RecipientOrganizationInn?: string | null;
  RecipientOrganizationKpp?: string | null;
  RecipientOrganizationName?: string | null;
  RegionalProductSignCanEdit?: boolean;
  RegionalProductSignChecked?: boolean;
  RegionalProductSignTenantId?: number | null;
  Regions?: Array<ApiRegionKladrDto> | null;
  RequestDateEnd?: string | null;
  RequestDateStart?: string | null;
  SelfDeliveryPlace?: string | null;
  SellerCode?: string | null;
  ShortOkeiName?: string | null;
  Signature?: string | null;
  Tags?: Array<ApiTagViewDto> | null;
  TruType?: ApiTruTypeEnum | null;
  UnitPrice?: number | null;
  UnitPriceInCurrency?: number | null;
  UsedClassificatorType?: ApiUsedClassificatorTypeEnum | null;
  UsedClassificatorTypeDescription?: string | null;
  VatRate?: number | null;
  VatRateDecimal?: number | null;
  VatRateDescription?: string | null;
  VatRateState?: ApiVatRateStateEnum;
  VendorCode?: string | null;
  WholesalePercentage?: number | null;
}